refactor: migrate dlna_proxy to party-display-core
