+++ app/src-tauri/party-display-core/Cargo.toml
@@ -7,6 +7,6 @@ edition = "2021"
 serde     = { version = "1", features = ["derive"] }
 serde_json = "1"
 roxmltree  = "0.20"
-reqwest    = { version = "0.12", features = ["json"] }
+reqwest    = { version = "0.12", features = ["json", "stream"] }
 tokio      = { version = "1", features = ["net", "io-util"] }
 futures    = "0.3"
