+++ app/src-tauri/party-display-core/src/lib.rs
@@ -1,0 +1,1 @@
+pub mod dlna_proxy;
